api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
psycopg[binary]==3.2.13
python-dotenv==1.1.1
requests==2.32.5
email-validator==2.2.0
